refactor(auth): extract shared JWT + atomic-write helpers into auth/shared.rs
Several credential helpers were provider-agnostic but lived in codex_auth.rs,
and the xAI, Claude, and Cursor stores reached across into them. Move them into
a new `auth/shared.rs` so no provider auth module depends on a sibling:

  jwt_claims, jwt_exp, is_token_valid_at   (JWT expiry parsing)
  write_auth_file_atomic + write_private + set_private_permissions (atomic
                                            private-file writeback)
  format_iso8601 + civil_from_days          (ISO-8601 timestamp)
  EXPIRY_BUFFER

codex_auth.rs keeps its ChatGPT/Codex-specific logic (jwt_account_id, the auth
store, refresh) and now imports the shared helpers. Consumers (xai_auth,
xai_login, claude_auth, claude_store, cursor_auth) import from `auth::shared`.

Pure move-refactor: no behavior, wire-format, credential-file, or public config
change. Each symbol keeps its original visibility. The moved helpers stay
covered by the existing codex/xai/claude/cursor tests that exercise them; the
codex test module gains the imports it needs (jwt_exp, base64, Duration).

Prerequisite for #81 (per-provider auth submodules): removes the
xai/claude -> codex cross-imports that would otherwise block the split.

Refs #80
