+++ src/auth/xai_login.rs
@@ -15,7 +15,7 @@ use anyhow::{anyhow, bail, Context};
 use serde_json::Value;
 use tokio::time::{sleep, Instant};
 
-use crate::auth::codex_auth::jwt_exp;
+use crate::auth::shared::jwt_exp;
 use crate::auth::default_xai_auth_path;
 use crate::auth::xai_auth::{
     parse_token_response, write_tokens, CLIENT_ID, DEVICE_CODE_GRANT_TYPE, DEVICE_CODE_URL, SCOPE,
@@ -93,7 +93,7 @@ pub async fn run(provider: &str) -> anyhow::Result<()> {
     if let Some(exp) = jwt_exp(&tokens.access_token) {
         println!(
             "Access token valid until {} (shunt refreshes it automatically).",
-            crate::auth::codex_auth::format_iso8601(exp)
+            crate::auth::shared::format_iso8601(exp)
         );
     }
     Ok(())
